Add `CATEGORIES` option to sorters and delete therefore unused function `sort_regions_in_reading_order`
